UBERF-6540: Fix isIndexable and clean wrong indexed documents
